third_party/WebKit/LayoutTests/webaudio/audiochannelmerger-basic.html

<!DOCTYPE html>
<html>

<head>
  <script src="../resources/js-test.js"></script>
  <script src="resources/compatibility.js"></script>
  <script src="resources/audio-testing.js"></script>
</head>

<body>
  <script>
    description("Basic tests for ChannelMerger.");
    window.jsTestIsAsync = true;

    var audit = Audit.createTaskRunner();

    // Task: Checking constraints in ChannelMergerNode.
    audit.defineTask('exceptions-channels', function (done) {
      var context = new OfflineAudioContext(2, 128, 44100);
      var merger;

      Should.notThrow(function () {
        merger = context.createChannelMerger();
      });

      Should.throwWithType('IndexSizeError', function () {
        merger = context.createChannelMerger(0);
      });

      Should.notThrow(function () {
        merger = context.createChannelMerger(32);
      });

      // Can't create a channel merger with 33 channels because the audio
      // context has a 32-channel-limit in Chrome.
      Should.throwWithType('IndexSizeError', function () {
        merger = context.createChannelMerger(33);
      });

      done();
    });

    // Task: checking the channel-related properties have the correct value and 
    // can't be changed.
    audit.defineTask('exceptions-properties', function (done) {
      var context = new OfflineAudioContext(2, 128, 44100);
      var merger = context.createChannelMerger();

      Should.beEqualTo('merger.channelCount', merger.channelCount, 1);

      Should.throwWithType('InvalidStateError', function () {
        merger.channelCount = 3;
      });

      Should.beEqualTo(
        'merger.channelCountMode', merger.channelCountMode, 'explicit'
      );

      Should.throwWithType('InvalidStateError', function () {
        merger.channelCountMode = 'max';
      });

      done();
    });

    audit.defineTask('finish', function (done) {
      finishJSTest();
      done();
    });

    audit.runTasks(
      'exceptions-channels',
      'exceptions-properties',
      'finish'
    );

    successfullyParsed = true;
  </script>
</body>

</html>
