third_party/WebKit/LayoutTests/webaudio/audiochannelmerger-basic-expected.txt

Basic tests for ChannelMerger.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "merger = context.createChannelMerger();" did not throw exception.
PASS "merger = context.createChannelMerger(0);" threw exception IndexSizeError.
PASS "merger = context.createChannelMerger(32);" did not throw exception.
PASS "merger = context.createChannelMerger(33);" threw exception IndexSizeError.
PASS merger.channelCount is equal to 1.
PASS "merger.channelCount = 3;" threw exception InvalidStateError.
PASS merger.channelCountMode is equal to explicit.
PASS "merger.channelCountMode = 'max';" threw exception InvalidStateError.
PASS successfullyParsed is true

TEST COMPLETE
